requirements.txt
flask==2.0.1
tensorflow==2.8.0
Pillow==8.3.2
numpy==1.21.3
matplotlib==3.5.1
opencv-python==4.5.5.64
protobuf==3.19.6
keras-preprocessing>=1.1.1
tensorflow-io-gcs-filesystem>=0.23.1
werkzeug==2.0.3